fix: fix AgentSnapper import
These changes fix the way AgentSnapper was imported and move the
code away from __init__.py into charmed_agent.py.
